Utils as a separate artefact References: #44
