+++ rx/src/main/kotlin/com/motorro/rxlcemodel/rx/LceModel.kt
@@ -18,6 +18,7 @@ import com.motorro.rxlcemodel.rx.service.CacheService
 import com.motorro.rxlcemodel.rx.service.NetService
 import com.motorro.rxlcemodel.rx.service.ServiceSet
 import com.motorro.rxlcemodel.rx.service.UpdatingServiceSet
+import com.motorro.rxlcemodel.utils.Logger
 import io.reactivex.rxjava3.core.Completable
 import io.reactivex.rxjava3.core.Observable
 import io.reactivex.rxjava3.core.Scheduler
