+++ rx/src/main/kotlin/com/motorro/rxlcemodel/rx/UpdatingLceModelWrapper.kt
@@ -14,6 +14,7 @@
 package com.motorro.rxlcemodel.rx
 
 import com.motorro.rxlcemodel.rx.service.UpdatingServiceSet
+import com.motorro.rxlcemodel.utils.Logger
 import io.reactivex.rxjava3.core.Completable
 import io.reactivex.rxjava3.core.Scheduler
