+++ settings.gradle
@@ -3,3 +3,4 @@ include ':sample', ':rx', ':disklrucache', ':kserializer'
 include ':viewmodel'
 include ':composeview'
 include ':lce'
+include ':utils'
